Refactor sample directory
