+++ sample/constraint/n_hot.py
